Move ExtraSensor to sensor file
